Phase 14: plans, quotas and IOC write endpoints
方案/訂閱/配額與 IOC 寫入端點(執行單 docs/spec/phases/phase-14.md)。

- migration V28(plans/subscriptions/import_jobs + ingestion_rejections.import_job_id)、
  V29(四個方案種子 + subscription:read 權限,三處同步:規格 §10.3、seed、RbacMatrix)
- Subscription 聚合(B1-B5)與 QuotaService 單一判定點:§10.6 的 14 個配額維度全部讀 plans 表,
  §9.7 的三種超限語意各有出口(429 / 403 / 413 / 分頁夾值)
- IOC 寫入:POST /iocs、POST /iocs/import(202 + jobId 非同步)、GET /iocs/import/{jobId}、
  POST /iocs/{id}/report-false-positive;ManualSubmissionAdapter 複用同一條 pipeline
- 前端 /iocs/new、/iocs/import、/settings/subscription

實作回饋修訂 10 項(ADR 0023、規格 00 §0.20),其中三項是照原規格實作會留下的洞:
ioc:publish 光做擁有權轉移仍然沒有任何人看得到、匯入不計入每日配額則每日上限可被繞過、
配額 0 回 429 等於告訴 client 一個永遠不會到來的重試時機。

判準:test -Ptest-integration -Dtest='QuotaEnforcementTest,ManualSubmissionTest,
FalsePositiveReportTest' 24/24;clean verify -Ptest-integration 644 tests;
frontend tsc/eslint/build/test(115)/format:check;dod.sh mvp 38/38。

+++ backend/ctip-core/src/main/java/com/ctip/application/stix/StixExportLimitExceededException.java
@@ -6,7 +6,7 @@ public class StixExportLimitExceededException extends RuntimeException {
     @java.io.Serial
     private static final long serialVersionUID = 1L;
 
-    public StixExportLimitExceededException(int maxObjects) {
+    public StixExportLimitExceededException(long maxObjects) {
         super("bundle 物件數超過上限 " + maxObjects + "(PLAN_LIMIT_EXCEEDED)");
     }
 }
